Completed problem 32 Refactor: solve now returns BigInteger

+++ src/main/java/it/mrbigs/Problem.java
@@ -1,8 +1,10 @@
 package it.mrbigs;
 
+import java.math.BigInteger;
+
 /**
  * Created by Simone on 26/01/2015.
  */
 public interface Problem {
-    public long solve();
+    public BigInteger solve();
 }
