+++ src/main/java/it/mrbigs/Problem10.java
@@ -2,6 +2,7 @@ package it.mrbigs;
 
 import it.mrbigs.util.PrimeChecker;
 
+import java.math.BigInteger;
 import java.util.LinkedList;
 import java.util.List;
 
@@ -14,7 +15,7 @@ public class Problem10 implements Problem {
     private static final long LIMIT = 2000000;
 
     @Override
-    public long solve() {
+    public BigInteger solve() {
         long primeSum = 0;
         List<Long> primes = new LinkedList<>();
         for (long number = 0; number < LIMIT; number++) {
@@ -25,13 +26,13 @@ public class Problem10 implements Problem {
                 }
             } catch (Exception e) {
                 System.out.println("Problem10::solve() --> PrimeChecker::call() " + e.getMessage());
-                return -1;
+                return BigInteger.valueOf(-1);
             }
             if (number % 10000 == 0) {
                 System.out.println(number);
             }
         }
 
-        return primeSum;
+        return BigInteger.valueOf(primeSum);
     }
 }
