+++ src/main/java/it/mrbigs/Problem12.java
@@ -2,6 +2,8 @@ package it.mrbigs;
 
 import it.mrbigs.util.Common;
 
+import java.math.BigInteger;
+
 /**
  * Created by Simone on 31/01/2015.
  */
@@ -10,7 +12,7 @@ public class Problem12 implements Problem {
     private static int LIMIT = 500;
 
     @Override
-    public long solve() {
+    public BigInteger solve() {
         long currentNumber = 0;
         int numberOfDivisors = 0;
         for (int i = 1; numberOfDivisors <= LIMIT; i++) {
@@ -20,6 +22,6 @@ public class Problem12 implements Problem {
                 numberOfDivisors = numberOfDivisors * (exponent + 1);
             }
         }
-        return currentNumber;
+        return BigInteger.valueOf(currentNumber);
     }
 }
