+++ src/main/java/it/mrbigs/Problem15.java
@@ -2,6 +2,8 @@ package it.mrbigs;
 
 import it.mrbigs.util.Common;
 
+import java.math.BigInteger;
+
 /**
  * Created by Simone on 08/02/2015.
  */
@@ -10,9 +12,9 @@ public class Problem15 implements Problem {
     private static final long N = 20;
 
     @Override
-    public long solve() {
+    public BigInteger solve() {
 
-        return Common.factorial(N * 2).divide(Common.factorial(N).multiply(Common.factorial(N))).longValue();
+        return Common.factorial(N * 2).divide(Common.factorial(N).multiply(Common.factorial(N)));
     }
 
 }
