+++ src/main/java/it/mrbigs/Problem20.java
@@ -2,6 +2,8 @@ package it.mrbigs;
 
 import it.mrbigs.util.Common;
 
+import java.math.BigInteger;
+
 /**
  * Created by Simone on 09/02/2015.
  */
@@ -11,12 +13,12 @@ public class Problem20 implements Problem {
     private static final long NUMBER = 100;
 
     @Override
-    public long solve() {
+    public BigInteger solve() {
         String stringFact = Common.factorial(NUMBER).toString();
         long sum = 0;
         for (char c : stringFact.toCharArray()) {
             sum += Integer.parseInt(String.valueOf(c));
         }
-        return sum;
+        return BigInteger.valueOf(sum);
     }
 }
