+++ src/main/java/it/mrbigs/Problem3.java
@@ -2,6 +2,8 @@ package it.mrbigs;
 
 import it.mrbigs.util.Common;
 
+import java.math.BigInteger;
+
 /**
  * Created by Simone on 26/01/2015.
  */
@@ -12,14 +14,14 @@ public class Problem3 implements Problem {
 //    private long dividend = 13195L;
 
     @Override
-    public long solve() {
+    public BigInteger solve() {
 
         for (Integer factor : Common.primeFactors(dividend).keySet()) {
             if(factor > maxFactor) {
                 maxFactor = factor;
             }
         }
-        return maxFactor;
+        return BigInteger.valueOf(maxFactor);
     }
 
 }
