+++ src/main/java/it/mrbigs/Problem4.java
@@ -2,12 +2,14 @@ package it.mrbigs;
 
 import it.mrbigs.util.Common;
 
+import java.math.BigInteger;
+
 /**
  * Created by Simone on 30/01/2015.
  */
 public class Problem4 implements Problem {
     @Override
-    public long solve() {
+    public BigInteger solve() {
         int max = 0;
         for (int i = 100; i <= 999; i++) {
             for (int j = 100; j <= 999; j++) {
@@ -16,6 +18,6 @@ public class Problem4 implements Problem {
                 }
             }
         }
-        return max;
+        return BigInteger.valueOf(max);
     }
 }
