+++ src/main/java/it/mrbigs/Problem6.java
@@ -1,5 +1,7 @@
 package it.mrbigs;
 
+import java.math.BigInteger;
+
 /**
  * Created by Simone on 30/01/2015.
  */
@@ -9,13 +11,13 @@ public class Problem6 implements Problem {
     private static final long LIMIT = 100;
 
     @Override
-    public long solve() {
+    public BigInteger solve() {
         int sumOfSquares = 0;
         int sum = 0;
         for (int i = 1; i <= LIMIT; i++) {
             sumOfSquares += i*i;
             sum += i;
         }
-        return sum*sum - sumOfSquares;
+        return BigInteger.valueOf(sum*sum - sumOfSquares);
     }
 }
