+++ src/main/java/it/mrbigs/Problem9.java
@@ -1,5 +1,7 @@
 package it.mrbigs;
 
+import java.math.BigInteger;
+
 /**
  * Created by Simone on 30/01/2015.
  */
@@ -8,16 +10,16 @@ public class Problem9 implements Problem {
     private static final int SUM = 1000;
 
     @Override
-    public long solve() {
+    public BigInteger solve() {
         for (int i = 0; i < SUM; i++) {
             for (int j = i+1; j < SUM; j++) {
                 for (int k = j+1; k < SUM; k++) {
                     if (i+j+k == 1000 && i*i + j*j == k*k) {
-                        return i*j*k;
+                        return BigInteger.valueOf(i*j*k);
                     }
                 }
             }
         }
-        return -1;
+        return BigInteger.valueOf(-1);
     }
 }
